requirements.txt
fastapi==0.115.5
httpx==0.27.2
openai==1.42.0
pillow==10.4.0
pydantic==2.9.2
python-dotenv==1.0.1
reportlab==4.2.2
requests==2.32.3
streamlit==1.37.1
uvicorn[standard]==0.30.6
